refactor: reorganize auth service files and update imports
