+++ src/definition/handlers/IHandler.ts
@@ -18,7 +18,6 @@ export interface IHandlerParams {
     persis: IPersistence;
     triggerId?: string;
     threadId?: string;
-    language?: string;
 }
 
 export interface IHandler {
@@ -31,7 +30,6 @@ export interface IHandler {
     persis: IPersistence;
     triggerId?: string;
     threadId?: string;
-    language?: string;
     Help(): Promise<void>;
     sendDefault(): Promise<void>;
     Login(): Promise<void>;
